AWS_Strands/QuickResearch_Quinten/requirements.txt
# AWS Strands Agent Dependencies
strands-agents
strands-agents-tools
python-dotenv
anthropic

#search
crawl4ai